requirements.txt
django==1.10.2
djangorestframework
social-auth-app-django
fcm-django
django-guardian
